consolidate fixtures for segmentation/filter tests
