+++ tests/modules/segmentation/filter/test_filter_metric.py
@@ -8,13 +8,13 @@ from ophys_etl.modules.segmentation.filter.roi_filter import (
 
 
 @pytest.fixture(scope='session')
-def img_fixture():
+def small_img_fixture():
     data = np.arange(1024, dtype=float).reshape(32, 32)
     return data
 
 
 @pytest.fixture(scope='session')
-def roi_list_fixture():
+def small_roi_list_fixture():
 
     roi_list = []
     roi = OphysROI(x0=0, y0=3, height=3, width=3,
@@ -58,19 +58,19 @@ def roi_list_fixture():
      ('median', 88.8, 130.0, set([0, 1, 2])),
      ('median', None, 130.0, set([0, 1, 2, 3]))]
 )
-def test_filter_on_stat(img_fixture, roi_list_fixture,
+def test_filter_on_stat(small_img_fixture, small_roi_list_fixture,
                         stat_name, min_val, max_val, expected_valid):
 
-    this_filter = ROIMetricStatFilter(img_fixture,
+    this_filter = ROIMetricStatFilter(small_img_fixture,
                                       stat_name,
                                       min_metric=min_val,
                                       max_metric=max_val)
 
-    result = this_filter.do_filtering(roi_list_fixture)
+    result = this_filter.do_filtering(small_roi_list_fixture)
     valid_set = set([roi.roi_id for roi in result['valid_roi']])
     invalid_set = set([roi.roi_id for roi in result['invalid_roi']])
     assert len(valid_set.intersection(invalid_set)) == 0
-    assert len(valid_set) + len(invalid_set) == len(roi_list_fixture)
+    assert len(valid_set) + len(invalid_set) == len(small_roi_list_fixture)
     assert valid_set == expected_valid
